remove dangling mod
